Much better flush handling.
Now, responding to flushes is quicker. There still is an extra
cycle in the registering of do_branch, we will have to see
if that's necessary. But at least now requests are restarted
from the new target address as soon as possible, even while
flushing is happening. The trick is that requests are stopped
for one clock cycle due to a branch, partly to ensure that
we handle out-of-page branches properly but partly to force
a one-cycle drop of response.ready, which marks the end
of the flush.