New reg_file interface.
Here we introduce the ready/valid based interfaces to the reg_file.
The latency is constant 1 clock cycle (even for bypass) and
back-pressure is (hopefully) properly handled from the response
interface. I reworked the reservation logic as well. Not only it is
easier to read, hopefully it's also better for synthesis.

NOTE: the 1 clock cycle latency is somewhat sub-optimal for silicon
where we implement memories with registers, but works well for FPGAs.
It also isn't all that big of a deal as the output register can be
incorporated into the registering at the output of 'decode' not
resulting in a two-cycle decode latency.