An almost complete re-write of Execute.
This one shows some minimal signs of life,
but more importantly it synthesizes to
something that closes timing >100MHz
on a MAX10.

There are two additional improvements here:

1. A way to drive the full Quartus flow from
  within Python (synth.py)
2. A fairly automated way of wrapping an IP
  in a shift-register to remove dependency
  on I/O pincount to do synthesis runs.
  This module (in scan.py) could later easily
  be converted into a JTAG scan-chain implementation.

There's also a pretty complete (though rather buggy)
re-write of memory here as well. I'm sure it's not
working, but since Memory is now part of Execute,
it needs to be checked in at the same time.

This particular IP have seen 0 testing. All I know
is that it synthesizes into ... something.