Changes to get to timing closure at 100MHz.
Now, this is not *real* timing closure: I added some
registers that totally break the functionality just as an
experiment. However this pin-points the locations where
we somehow need to be break large logic cones. These registers
are added through 'TIMING_CLOSURE_REG', so they are easy
to find or even turn off.