In fact, the first instruction went through the pipeline!!!
Albeit incorrectly as the RF is uninitialized. This is a simulation
issue, because None ^ None is None, even though we know that
the two inputs are the same (the same None as it were) so the
result should be 0. However, the simulator doesn't (and can't)
know this. In other words, I'll have to initialize the registers
somewhat differently. Luckily we have the 'tiny' load operations.