Pipelined jump works!
There were a bunch of problems, mostly in the RF not taking
do_branch into consideration in clearing the state.

It's nice that we actually tried to issue an NONE instruction
after the jump, so things were obviously going south.

The recovery of the branch is still painfully slow, but at
least it works.