Flushing of fetch was wrong.
Interesting that so far it didn't cause issues.
At any rate, apparently the bus IF does have
some pipelining to it (it can accept a request
before it provides a response) which confused
the fetch flushing logic.

There's still about two clock-cycles worth of
latency that could be removed, but that'll
have to wait.