BusIF now supports 4 DRAM banks.
This - with the right pin-muxing - brings the processor
fully in sync with the pinout sketched out in the
documentation.

This final change enables 128kByte to 16MByte
memory sizes with by-4 chips earlier in time:
1982-based chips allow for 512k total and we
can reach 2MByte by 1986. We also never need
more than 16 chips to support memory sizes
which helps with loading of the ADDR/CAS/RAS
etc. drivers.