Yet another bus_if pin-muxing reorg.
This time, gone are the 4 DRAM banks, only 2 remained.
The address muxing is also straightforward: first 11 address bits
for column, next 11 for row.

What we do allow is to set the size of the banks from 16 to 22
address bits, and swapping of the banks in the address space.

This together allows for auto-detection of the RAM size and
organizing the banks such that both banks appear as a
contiguous chunk of address space, even if the two banks
are populated with different sized DRAMs.

This organization also simplifies non-DRAM interfacing as
all the interesting address bits for address decode appear
on the first address cycle.