stack-ops work.
There were several rather trivial problems with these
operations in the assembler and decode.

The brunt of the issue was that pipeline stalls were
handled incorrectly for load/stores in execute: they've
used the stage-1 operand, instead of stage-2. Plus, stores
should not stall the pipeline for the duration of the
bus operation (unless of course there's a second
memory op behind them), but loads should:
we don't support out-of-order finalization, so
even though the hazards are handled in decode
already, we need to make sure that load results
are stored before that of any subsequent operation.