Fixed fetch dropping responses on the floor.
This was a fun one: we occasionally request a new burst from
the bus interface, but haven't accounted for the fact that there
were some outstanding responses from our previous burst.

This meant that we could, under some rare circumstances, overflow
the instruction queue and (since there is no way to back-pressure DRAM)
dropped some responses on the floor.

While I was at it, added a comparator to make sure we never requested
more than 8 cycles of burst from the bus interface. This is to be good
citizens on the bus and don't hog it for too long.