No braking of bursts for in-page branches.
This change allows for fetch to continue with a burst if the branch
stayed within the same DRAM page. Since most branches supposed
to be in tight loops, this should improve branch mis-predict penalty
quite a bit. And it does.

I think for now, this is going to be the last perf-related change and
I'll concentrate on getting correctness checked out.