Fully putting trust in internal PLL
This approach embraces the use of the internal PLL.

We now only depend on the external 10MHz clock and
generate our internal 50MHz (and 10MHz as a matter of fact)
from this external clock.

The benefit of this is that Quartus is now able to derive all
CDC timings as the relationship of the clocks is known.

The downside of course is ... well ... depending on the internal PLL.

The design now closes timing for setup times, but NOT
for hold times. I still don't get how we can get hold violations,
even after staring at the timing report for a while.

There also seems to be some instability in the response to reset,
but I'm not not certain what that could be. There are several
possibilities, including no bound-removal, doggy sampling of
the reset pin, not resetting something that should reset
or even the aforementioned hold violations.

At any rate, the system now largely runs at speed.