Removed type load/stores added type-tests.
A new set of instructions are added to test
for expected register types. These can be used
in function prologs to ensure that argument types
are as expected. Specifically, they can be used to
test for the fact that arguments can be safely stored
in memory without the risk of overflow.