CSRs are now special instructions
This is a momentous change: I gave in and created special instructions
for CSR loads and stores. I also divided the address space up to allow
for scheduler-mode only CSRs.

The trigger for this change was that VEND **must not** be accessible
through user code, and I really didn't want to introduce privileged
instructions.

The other big change is that 'check_isa.py' started comparing instruction
details to summary. In fact it passes now, thus the hundreds of other
changes all over.